+++ example/websocket/client/async-ssl/websocket_client_async_ssl.cpp
@@ -47,7 +47,7 @@ class session : public std::enable_shared_from_this<session>
     tcp::resolver resolver_;
     websocket::stream<
         beast::ssl_stream<beast::tcp_stream>> ws_;
-    beast::multi_buffer buffer_;
+    beast::flat_buffer buffer_;
     std::string host_;
     std::string text_;