+++ example/websocket/server/async/websocket_server_async.cpp
@@ -44,7 +44,7 @@ fail(beast::error_code ec, char const* what)
 class session : public std::enable_shared_from_this<session>
 {
     websocket::stream<beast::tcp_stream> ws_;
-    beast::multi_buffer buffer_;
+    beast::flat_buffer buffer_;
 
 public:
     // Take ownership of the socket