+++ example/websocket/server/stackless-ssl/websocket_server_stackless_ssl.cpp
@@ -52,7 +52,7 @@ class session
     , public std::enable_shared_from_this<session>
 {
     websocket::stream<beast::ssl_stream<beast::tcp_stream>> ws_;
-    beast::multi_buffer buffer_;
+    beast::flat_buffer buffer_;
 
 public:
     // Take ownership of the socket