+++ example/websocket/server/sync-ssl/websocket_server_sync_ssl.cpp
@@ -63,7 +63,7 @@ do_session(tcp::socket& socket, ssl::context& ctx)
         for(;;)
         {
             // This buffer will hold the incoming message
-            beast::multi_buffer buffer;
+            beast::flat_buffer buffer;
 
             // Read a message
             ws.read(buffer);