+++ example/websocket/server/sync/websocket_server_sync.cpp
@@ -54,7 +54,7 @@ do_session(tcp::socket& socket)
         for(;;)
         {
             // This buffer will hold the incoming message
-            beast::multi_buffer buffer;
+            beast::flat_buffer buffer;
 
             // Read a message
             ws.read(buffer);