+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 225:
 
 * Tidy up an unused function
 * Fix wsload jamfile
+* Examples use flat_buffer
 
 --------------------------------------------------------------------------------